requirements/requirement_sam.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
exceptiongroup>=1.0.0
executing==0.8.3
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
idna==3.11
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.1.6
mcp==1.24.0
ml_dtypes==0.5.4
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
# nvidia-cufile-cu12: Linux-only (no Windows wheel); omit on Windows
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvshmem-cu12==3.3.20  # Linux-only; uncomment on Linux if needed
nvidia-nvtx-cu12==12.8.90
onnx==1.20.0
onnxruntime==1.23.2
openai==2.11.0
opencv-python==4.12.0.88
packaging==24.2
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==12.0.0
platformdirs==4.3.7
prompt_toolkit==3.0.43
protobuf==6.33.2
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyzmq==26.2.0
referencing==0.37.0
rpds-py==0.30.0
# Editable install for segment_anything (install from utils/third_party/sam)
-e ./utils/third_party/sam
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
stack_data==0.2.0
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
# triton==3.5.1
typing-inspection==0.4.2
typing_extensions>=4.13.2
uvicorn==0.38.0
wcwidth==0.2.5
